ai-service/requirements.txt
--index-url https://download.pytorch.org/whl/cu118
torch 
torchvision 
torchaudio
git+https://github.com/m-bain/whisperX.git

numpy
pandas
opencv-python
transformers
langchain
openai
anthropic
datasets
scikit-learn
grpcio
grpcio-tools
protobuf
boto3
accelerate
sentence-transformers
spacy==3.8.3
tiktoken
openai-whisper
librosa
facenet-pytorch
pyannote.audio
ultralytics
pytesseract
Pillow